requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
ta>=0.11.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
flask>=3.0.0
psutil>=5.9.0
